+++ lib/src/lexer.h
@@ -21,11 +21,11 @@ typedef struct {
   TSInput input;
   TSLogger logger;
 
-  uint32_t included_range_count;
-  uint32_t current_included_range_index;
-  uint32_t chunk_start;
-  uint32_t chunk_size;
-  uint32_t lookahead_size;
+  uint64_t included_range_count;
+  uint64_t current_included_range_index;
+  uint64_t chunk_start;
+  uint64_t chunk_size;
+  uint64_t lookahead_size;
   bool did_get_column;
 
   char debug_buffer[TREE_SITTER_SERIALIZATION_BUFFER_SIZE];
@@ -36,11 +36,11 @@ void ts_lexer_delete(Lexer *);
 void ts_lexer_set_input(Lexer *, TSInput);
 void ts_lexer_reset(Lexer *, Length);
 void ts_lexer_start(Lexer *);
-void ts_lexer_finish(Lexer *, uint32_t *);
+void ts_lexer_finish(Lexer *, uint64_t *);
 void ts_lexer_advance_to_end(Lexer *);
 void ts_lexer_mark_end(Lexer *);
-bool ts_lexer_set_included_ranges(Lexer *self, const TSRange *ranges, uint32_t count);
-TSRange *ts_lexer_included_ranges(const Lexer *self, uint32_t *count);
+bool ts_lexer_set_included_ranges(Lexer *self, const TSRange *ranges, uint64_t count);
+TSRange *ts_lexer_included_ranges(const Lexer *self, uint64_t *count);
 
 #ifdef __cplusplus
 }
